+++ src/main/java/com/pedroavila/domain/service/dto/CreateBranchOfficeResult.java
@@ -2,5 +2,5 @@ package com.pedroavila.domain.service.dto;
 
 import java.util.Date;
 
-public record CreateBranchOfficeResult(int id, int companyId, int code, String name, String address, String phone, int status, Date creationDate) {
+public record CreateBranchOfficeResult(Long id, int companyId, int code, String name, String address, String phone, int status, Date creationDate) {
 }
